requirements.txt
streamlit
streamlit-drawable-canvas
numpy
pillow